Squashed commit of the following:

commit 927335a9d770afd3f976c5d433dfbf59c358f1d3
Author: russ <rucombs@cisco.com>
Date:   Sat Dec 1 09:08:48 2018 -0500

    build: clean up additional warnings

commit 237dc040cedf560fa2e10127c5db9cf6b313fd8c
Author: russ <rucombs@cisco.com>
Date:   Fri Nov 30 14:54:58 2018 -0500

    build: fix some int format specifiers

commit bd1b6a63ed81d99dcec9404bdf08a886d04acf48
Author: russ <rucombs@cisco.com>
Date:   Fri Nov 30 14:22:37 2018 -0500

    build: fix some int type conversion warnings

commit 489e5ce892199bef55974bbd1b85281fd1e1c453
Author: russ <rucombs@cisco.com>
Date:   Fri Nov 30 08:17:16 2018 -0500

    parameter: remove arbitrary one day limit on timers

commit dadfae41cc9e2cc73809688de593f93b9d4e6a20
Author: russ <rucombs@cisco.com>
Date:   Thu Nov 29 08:46:55 2018 -0500

    snort: add --help-limits to output max* values

commit 8886dbaec359e974b016d5f33e50fd75a62722ef
Author: russ <rucombs@cisco.com>
Date:   Thu Nov 29 07:50:07 2018 -0500

    parameter: add maxSZ upper bound for int sizes

commit 4d29c63de0c65315f3cfb7c30d63604366369095
Author: russ <rucombs@cisco.com>
Date:   Wed Nov 28 13:39:16 2018 -0500

    parameter: add max31, max32, and max53 for int upper bounds

commit 25c837adb622de44ea746d8081498b4a3c3e3d7a
Author: russ <rucombs@cisco.com>
Date:   Tue Nov 27 08:27:05 2018 -0500

    spelling: fix typo, disable checks on test data

commit acca70d032f3de0ca5e3ca0ef4e80dbd2199d743
Author: russ <rucombs@cisco.com>
Date:   Mon Nov 26 22:46:47 2018 -0500

    parameter: clean up some signed/unsigned mismatches

commit f0d5e1a033ccf16a57af6551f098bb43adf19fdd
Author: russ <rucombs@cisco.com>
Date:   Mon Nov 26 17:37:52 2018 -0500

    parameter: remove ineffective -1 from pcre_match_limit*

commit f52729f59324519be52a0058fbe3c104bb056783
Author: russ <rucombs@cisco.com>
Date:   Mon Nov 26 12:03:55 2018 -0500

    parameter: use bool instead of int for bools

commit c845621d7900f97da697a966b4b2fe8af9008614
Author: russ <rucombs@cisco.com>
Date:   Mon Nov 26 11:44:57 2018 -0500

    parameter: use consistent default port ranges

commit ab3be3c2121bb9a2f89f65ed3712eb5f5f68046b
Author: russ <rucombs@cisco.com>
Date:   Mon Nov 26 08:32:20 2018 -0500

    parameter: clean up upper bounds

commit 78aaf9fc08d6b05e8949426eace6a83e06f731ea
Author: russ <rucombs@cisco.com>
Date:   Fri Nov 23 10:52:57 2018 -0500

    parameter: build out validation unit tests

commit ed4235ab439d1966e645378f97fd079672329091
Author: russ <rucombs@cisco.com>
Date:   Fri Nov 23 09:16:35 2018 -0500

    perf_monitor: fix benign parameter errors

commit bfcbb47e5c01608532df3308cf2e684c5e06b6dc
Author: russ <rucombs@cisco.com>
Date:   Fri Nov 23 09:14:57 2018 -0500

    parameter: reorgranize for unit tests

+++ src/detection/fp_create.cc
@@ -64,11 +64,11 @@ static const char* s_group = "";
 static void fpDeletePMX(void* data);
 
 static int fpGetFinalPattern(
-    FastPatternConfig*, PatternMatchData*, const char*& ret_pattern, int& ret_bytes);
+    FastPatternConfig*, PatternMatchData*, const char*& ret_pattern, unsigned& ret_bytes);
 
 static void print_nfp_info(const char*, OptTreeNode*);
 static void print_fp_info(const char*, const OptTreeNode*, const PatternMatchData*,
-    const char* pattern, int pattern_length);
+    const char* pattern, unsigned pattern_length);
 
 static int finalize_detection_option_tree(SnortConfig* sc, detection_option_tree_root_t* root)
 {
@@ -404,7 +404,7 @@ static int fpFinishPortGroupRule(
         pg->add_rule();
 
     const char* pattern;
-    int pattern_length;
+    unsigned pattern_length;
 
     if (fpGetFinalPattern(fp, pmd, pattern, pattern_length) == -1)
         return -1;
@@ -755,7 +755,7 @@ static void fpFreeRuleMaps(SnortConfig* sc)
 
 static int fpGetFinalPattern(
     FastPatternConfig* fp, PatternMatchData* pmd,
-    const char*& ret_pattern, int& ret_bytes)
+    const char*& ret_pattern, unsigned& ret_bytes)
 {
     if ( !fp or !pmd )
     {
@@ -763,7 +763,7 @@ static int fpGetFinalPattern(
     }
 
     const char* pattern = pmd->pattern_buf;
-    int bytes = pmd->pattern_size;
+    unsigned bytes = pmd->pattern_size;
 
     // Don't mess with:
     //
@@ -806,10 +806,9 @@ static int fpGetFinalPattern(
          * beyond state 0 as long as the next input char is 0x00 */
         if ( fp->get_trim() )
         {
-            bytes =
-                flp_trim(pmd->pattern_buf, pmd->pattern_size, &pattern);
+            bytes = flp_trim(pmd->pattern_buf, pmd->pattern_size, &pattern);
 
-            if (bytes < (int)pmd->pattern_size)
+            if (bytes < pmd->pattern_size)
             {
                 // The pattern is all '\0' - use the whole pattern. This potentially
                 // hurts the performance boost gained by stripping leading zeros.
@@ -1637,8 +1636,9 @@ void get_pattern_info(const PatternMatchData* pmd,
     opts += " )";
 }
 
-static void print_fp_info(const char* group, const OptTreeNode* otn, const PatternMatchData* pmd,
-    const char* pattern, int pattern_length)
+static void print_fp_info(
+    const char* group, const OptTreeNode* otn, const PatternMatchData* pmd,
+    const char* pattern, unsigned pattern_length)
 {
     std::string hex, txt, opts;
